test(): add  proxy.ts teest
