+++ .vite/deps/_metadata.json
@@ -1,0 +1,8 @@
+{
+  "hash": "470d54d6",
+  "configHash": "60fdab46",
+  "lockfileHash": "e3b0c442",
+  "browserHash": "424d5825",
+  "optimized": {},
+  "chunks": {}
+}
