--- background-dynamic-rules.js
@@ -1,26 +1,0 @@
-var rebuildRules = undefined;
-if (typeof chrome !== "undefined" && chrome.runtime && chrome.runtime.id) {
-    rebuildRules = async function (domain) {
-        const domains = [domain];
-        /** @type {chrome.declarativeNetRequest.Rule[]} */
-        const rules = [{
-            id: 1,
-            condition: {
-                urlFilter: '.*',
-                ignoreCase: true
-            },
-            action: {
-                type: 'modifyHeaders',
-                requestHeaders: [{
-                    header: 'origin',
-                    operation: 'set',
-                    value: `http://${domain}`,
-                }],
-            },
-        }];
-        await chrome.declarativeNetRequest.updateDynamicRules({
-            removeRuleIds: rules.map(r => r.id),
-            addRules: rules,
-        });
-    }
-}
